fix(after selfReview): update fields from double to BigDecimal

+++ src/main/java/maxim/butenko/dto/ExchangeRateDTO.java
@@ -2,11 +2,14 @@ package maxim.butenko.dto;
 
 import lombok.Builder;
 import lombok.Value;
+
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class ExchangeRateDTO {
     long id;
     CurrencyDTO baseCurrencyId;
     CurrencyDTO targetCurrencyId;
-    double rate;
+    BigDecimal rate;
 }
